Factor out all metric processing worker functions
- what remains is a short  processMetrics() function
- btw, fix issue with double-slash separated custom metrics in heapster: expose unprefixed namespace as custom metric name
